% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ap_config_path}
\alias{ap_config_path}
\title{Return the user AutoPlotR config path}
\usage{
ap_config_path(filename = "config.yml")
}
\arguments{
\item{filename}{Config filename.}
}
\value{
Path to an AutoPlotR user config file.
}
\description{
Return the user AutoPlotR config path
}
\examples{
ap_config_path()
}
