% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{ap_load_viz_rules}
\alias{ap_load_viz_rules}
\title{Load visualization rules}
\usage{
ap_load_viz_rules(plot_type = NULL, refresh = FALSE)
}
\arguments{
\item{plot_type}{Optional plot type such as \code{"scatter"} or \code{"histogram"}.}

\item{refresh}{Whether to refresh the internal rules cache.}
}
\value{
A list of visualization rules.
}
\description{
Load visualization rules
}
\examples{
names(ap_load_viz_rules())
ap_load_viz_rules("scatter")$plot_type
}
