% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{ap_viz_rules_path}
\alias{ap_viz_rules_path}
\title{Return the visualization rule path}
\usage{
ap_viz_rules_path(scope = c("package", "user", "project"))
}
\arguments{
\item{scope}{One of \code{"package"}, \code{"user"}, or \code{"project"}.}
}
\value{
Path to a visualization rules file.
}
\description{
Return the visualization rule path
}
\examples{
ap_viz_rules_path()
}
