% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{print.ap_live_runtime}
\alias{print.ap_live_runtime}
\title{Print an AutoPlotR live runtime}
\usage{
\method{print}{ap_live_runtime}(x, ...)
}
\arguments{
\item{x}{An \code{ap_live_runtime}.}

\item{...}{Unused.}
}
\description{
Print an AutoPlotR live runtime
}
