% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print.ap_plot_result}
\alias{print.ap_plot_result}
\title{Print an AutoPlotR plot result}
\usage{
\method{print}{ap_plot_result}(x, ...)
}
\arguments{
\item{x}{An \code{ap_plot_result}.}

\item{...}{Unused.}
}
\description{
Print an AutoPlotR plot result
}
