% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print.ap_viz_plan}
\alias{print.ap_viz_plan}
\title{Print an AutoPlotR visualization plan}
\usage{
\method{print}{ap_viz_plan}(x, ...)
}
\arguments{
\item{x}{An \code{ap_viz_plan}.}

\item{...}{Unused.}
}
\description{
Print an AutoPlotR visualization plan
}
